BookCatalog/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
gunicorn==21.2.0
sqlparse==0.5.3
tzdata==2025.2
